Check if docker machine is running before killing it.

+++ kill.sh
@@ -5,12 +5,16 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 chmod +x $DIR/config.sh
 . $DIR/config.sh
 
-#Check if forced is required
-if [ $# == 1 ] && [ "$@" == "-f" ]
+#Stopping docker-machine
+if [ "$( docker-machine status "$DOCKER_MACHINE_NAME" )" == "Running" ]
 then
-	echo "Forcefully stopping virtual-machine"
-	VBoxManage controlvm "$DOCKER_MACHINE_NAME" poweroff
-else
-	echo "Stopping docker-machine"
-	docker-machine stop "$DOCKER_MACHINE_NAME"
+	#Check if forced is required
+	if [ $# == 1 ] && [ "$@" == "-f" ]
+	then
+		echo "Forcefully stopping virtual-machine"
+		VBoxManage controlvm "$DOCKER_MACHINE_NAME" poweroff
+	else
+		echo "Stopping docker-machine"
+		docker-machine stop "$DOCKER_MACHINE_NAME"
+	fi
 fi
